Refactor utils and add minutes sport check
